feat: 사용자 삭제 시 연관 데이터 함께 삭제기능

+++ src/main/java/com/sprint/mission/discodeit/service/basic/BasicUserService.java
@@ -105,6 +105,16 @@ public class BasicUserService implements UserService {
     @Override
     public void delete(UUID id) {
 
+        User user = userRepository.findById(id).orElse(null);
+
+        if (user == null) {
+            return;
+        }
+
+        if (user.getProfileId() != null) {
+            binaryContentRepository.deleteById(user.getProfileId());
+        }
+        userStatusRepository.deleteByUserId(id);
         userRepository.deleteById(id);
     }
 }
